# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Arabic Game Translations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Mod {
  id: number;
  title: string;
  gameName: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  extraImages?: string[] | null;
  downloadCount: number;
  viewCount: number;
  /** @nullable */
  download1Label?: string | null;
  /** @nullable */
  download1Url?: string | null;
  /** @nullable */
  download2Label?: string | null;
  /** @nullable */
  download2Url?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ModInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  gameName: string;
  description?: string;
  imageUrl?: string;
  extraImages?: string[];
  download1Label?: string;
  download1Url?: string;
  download2Label?: string;
  download2Url?: string;
}

export interface ModUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  gameName?: string;
  description?: string;
  imageUrl?: string;
  extraImages?: string[];
  download1Label?: string;
  download1Url?: string;
  download2Label?: string;
  download2Url?: string;
}

export interface AdminStatus {
  isSetup: boolean;
}

export interface SetupAdminBody {
  /** @minLength 1 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface VerifyAdminBody {
  username: string;
  password: string;
}

export interface AdminVerifyResult {
  success: boolean;
}

export interface SiteStats {
  totalMods: number;
  totalDownloads: number;
  totalViews: number;
}

export interface Ad {
  id: number;
  /** @nullable */
  title?: string | null;
  imageUrl: string;
  linkUrl: string;
  position: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type AdInputPosition = typeof AdInputPosition[keyof typeof AdInputPosition];


export const AdInputPosition = {
  home: 'home',
  mod_detail: 'mod_detail',
} as const;

export interface AdInput {
  title?: string;
  imageUrl: string;
  linkUrl: string;
  position: AdInputPosition;
  isActive?: boolean;
}