# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Arabic Game Translations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ads:
    get:
      operationId: listAds
      tags: [admin]
      summary: List all ads
      responses:
        "200":
          description: List of ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      operationId: createAd
      tags: [admin]
      summary: Create ad (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
        "401":
          description: Unauthorized
  /ads/{id}:
    put:
      operationId: updateAd
      tags: [admin]
      summary: Update ad (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteAd
      tags: [admin]
      summary: Delete ad (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
components:
  schemas:
...
    Ad:
      type: object
      required: [id, imageUrl, linkUrl, position, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: string
        position:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    AdInput:
      type: object
      required: [imageUrl, linkUrl, position]
      properties:
        title:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        position:
          type: string
          enum: [home, mod_detail]
        isActive:
          type: boolean